+++ projects/rocshmem/CHANGELOG.md
@@ -2,6 +2,8 @@
 ## Unreleased - rocSHMEM 3.3.0 for ROCm x.x.x
 ### Added
 * Added new APIs:
+   * `rocshmem_quiet_on_stream`
+   * `rocshmem_sync_all_on_stream`
    * `rocshmem_TYPENAME_alltoall_wg`
    * `rocshmem_TYPENAME_alltoallv_wg`
    * `rocshmem_team_my_pe`
@@ -10,7 +12,7 @@
    * `rocshmem_barrier_wave`
    * `rocshmem_barrier_wg`
 * Added `ROCSHMEM_TEAM_WORLD` for the device code
-* Added new Enviroment variables:
+* Added new environment variables:
   * `OVERRIDE_NIC_FIRMWARE_CHECK`
   * `ROCSHMEM_GDA_NUM_QPS_PER_PE_DEFAULT_CTX`
   * `ROCSHMEM_GDA_NUM_QPS_PER_PE_USR_CTX`
@@ -65,7 +67,7 @@
    a shared memory region when the IPC transport is available to reach that region.
    Previously, it would return a null pointer.
 * `ROCSHMEM_RO_DISABLE_IPC` was renamed to `ROCSHMEM_DISABLE_MIXED_IPC`.
-  This enviroment variable was not documented for prior releases.
+  This environment variable was not documented for prior releases.
   It is now documented to inform users who were using this undocumented feature.
 
 ### Removed